requirements.txt
flask==2.3.3
gunicorn==22.0.0
cryptography==43.0.1
numpy==1.26.4
pycryptodome==3.20.0
werkzeug==2.3.8
waitress==3.0.1